Adds number prefix to order of Koans
All the koans start with 'About' anyways, much easier to predictably navigate if all the files are numbered in the order you'll be completing them
