+++ .tours/introduction.tour
@@ -1,4 +1,5 @@
 {
+  "$schema": "https://aka.ms/codetour-schema",
   "title": "Introduction",
   "steps": [
     {
@@ -8,7 +9,7 @@
     {
       "title": "Koans Location",
       "directory": "Koans",
-      "description": "# Koans Location\n\nAll of the code that you will be working with is located in the Koans directory. Each file represents an individual topic in the Path to Enlightenment.\n\n[**Next →**][#3]"
+      "description": "# Koans Location\n\nAll of the code that you will be working with is located in the Koans directory. Each file represents an individual topic in the Path to Enlightenment.\n\nThe file names are prefixed with a number representing the order they will be evaluated.\n\n[**Next →**][#3]"
     },
     {
       "title": "Path to Enlightenment",
@@ -24,13 +25,13 @@
     },
     {
       "title": "About Asserts",
-      "file": "Koans/AboutAsserts.cs",
+      "file": "Koans/1_AboutAsserts.cs",
       "line": 6,
       "description": "# About Asserts\n\nThis is the AboutAsserts Koan.\n\n[**Next →**][#6]"
     },
     {
       "title": "About Asserts",
-      "file": "Koans/AboutAsserts.cs",
+      "file": "Koans/1_AboutAsserts.cs",
       "line": 14,
       "description": "# About Asserts\n\nThe first challenge you will face is located here. Based on just looking at the highlighted code, can you figure out how to make it work?\n\n[**Next →**][#7]",
       "selection": {
@@ -46,7 +47,7 @@
     },
     {
       "title": "About Asserts",
-      "file": "Koans/AboutAsserts.cs",
+      "file": "Koans/1_AboutAsserts.cs",
       "line": 44,
       "description": "# About Asserts\n\nSometimes, you will see the value FILL_ME_IN. To solve the Koan, you must change the value.\n\n[**Next →**][#8]",
       "selection": {
