+++ Koans/10_AboutEnumerations.cs
