+++ Koans/11_AboutDictionary.cs
