+++ Koans/12_AboutClasses.cs
