+++ Koans/13_AboutInheritance.cs
