+++ Koans/17_AboutExceptions.cs
