+++ Koans/18_AboutGenericContainers.cs
