+++ Koans/19_AboutDelegates.cs
