+++ Koans/1_AboutAsserts.cs
