+++ Koans/20_AboutLambdas.cs
