+++ Koans/21_AboutDirectory.cs
