+++ Koans/22_AboutFile.cs
