+++ Koans/24_AboutBitwiseAndShiftOperator.cs
