+++ Koans/25_AboutDestructuring.cs
