+++ Koans/26_AboutPatternMatching.cs
