+++ Koans/27_AboutTuples.cs
