+++ Koans/28_AboutGlobalization.cs
