+++ Koans/29_AboutDisposable.cs
