+++ Koans/2_AboutBooleans.cs
