+++ Koans/30_AboutThreads.cs
