+++ Koans/3_AboutStrings.cs
