+++ Koans/5_AboutDecimals.cs
