+++ Koans/6_AboutNull.cs
