+++ Koans/7_AboutConstants.cs
