+++ Koans/9_AboutArrayAssignment.cs
